Enable Travis publishing of gem to rubygems.org
This contains the encrypted API key for the gds-paas rubygems.org user
(creds in the high cred store). This has been encrypted as described in
the Travis docs[1].

This is configured to only publish on git tags. This will result in a
workflow where we tag master when we're ready to publish a version.

It's necessary to pin to a specific ruby version, otherwise Travis will
attempt to publish multiple times for each tag push (once for each ruby
version sub-build).

[1]https://docs.travis-ci.com/user/deployment/rubygems/